Create a new Santa operating mode (standalone)
This adds a new operating mode to Santa called standalone mode.

When running in standalone mode TouchID can be used to approved binaries. If a
binary is properly signed a SigningID rule is generated otherwise a SHA256 rule
is generated. Note this lacks a GUI for browsing rules in the local rule db.