cleanup legacy components (#7)
This CL removes `com.northpolesec.santad.plist`. Prior to this change,
it was used to launch `com.northpolesec.santa.daemon` as a non-sysx.
Running as a non-sysx, it supported the kext -> sysx transition and then
relaunched itself as a sysx.

We don't need to support kext -> sysx transitions anymore, so let's
remove it and simply launch as a sysx from the start.

The change to have `-requestNeedsUserApproval:` `exit(1)` is intentional
and important. It keeps the install.sh and the pkg from blocking if the
sysx is not yet approved. This non-blocking behavior was previously
enjoyed because loading `com.northpolesec.santad.plist` is non-blocking,
any subsequent blocking calls to load the sysx were masked.

This CL also removes some code that cleaned up cruft from early versions
of Santa. This is also not needed anymore.