installer: support side-by-side install of NPS Santa [1/2] (#19)
This CL changes the installer to lay down Santa on disk as
Santa_NPS.app. The post install or daemon will rename Santa_NPS.app to
Santa.app when it is safe to do so.

A follow up CL will add logic in the daemon to idle while another
install of Santa is running.

To test the installer behavior as new label has been added
`package-dev`, which will package up a dev build. To support this,
`package.sh` is now a separate script that can be called for both
release and dev builds.

```
bazel build :package-dev -c opt
```