Address findings from migration testing (#91)
This PR adds additional cleanup for two cases:

1. Upgrading from NPS Santa to NPS Santa
The `com.northpolesec.santa.migration.plist` launch daemon artifact was
being left on disk. This was due to tamper protections preventing the
deletion on installation of the upgraded version. While benign, it's not
ideal to leave the artifact on disk. This file is now removed.

2. When Google Santa is attempted to be installed over NPS Santa, the
`com.google.santa.newsyslog.conf` artifact was not cleaned up.
During discussion, we decided it would be best that instead of always
attempting to remove this file, we should instead mv the contents to the
new `com.northpolesec.santa.newsyslog.conf` path. This will help NPS
deployments maintain the same configuration as the previous deployments
in the rare event any changes to this file were made.