config: services now watch for config changes (#121)
Previously we had limited watching for config changes to only
`com.northpolesec.santa.daemon`. This was because the other major
consumer, `santasyncservice`, received an empty config when the config
changed. It turns out this was due to the `NSUserDefaults`, by way of
the `SNTConfigurator` singleton, being initialized before
`santasyncservice` drops its root privileges.

This CL enables `santasyncservice`, and the other services, to watch for
config changes.

It also fixes the bug described above in `santasyncservice`,
`SNTConfigurator` is now initialized after dropping privileges.