standalone: Allow password fallback
This makes standalone mode usable in clamshell mode or on a desktop machine.

This change also modifies the authorization dialog for unsigned, non-bundled binaries to just show the binary name instead of the full path. In many cases the full path is very long and as the dialog is limited in width it makes the path squash up and become nearly unreadable.