Fix: Add signing ID checks for critical binaries (#164)
This PR adds signing ID checks for critical system binaries in addition
to the hash check.

In some scenarios e.g. a Santa upgrade in lockdown mode you may have two
versions of the critical binary that needs to run on the system. This
ensures that the later can run.

---------

Signed-off-by: Pete Markowsky <pmarkowsky@users.noreply.github.com>
Co-authored-by: Russell Hancox <russellhancox@users.noreply.github.com>