Restructure Watch Items objects to prepare for Proc FAA (#161)
This PR starts the process of breaking up `WatchItemPolicy` into
`DataWatchItemPolicy` and `ProcessWatchItemPolicy` to as a first step in
implementing Proc FAA.

`WatchItems` is also being broken up to abstract out the
`DataWatchItems` into their own object so that it can be operated on
singularly. A future PR will add `ProcessWatchItems` to work similarly.