Use a deque for the ring buffer for better dequeue performance (#182)
Replaces the underlying type of the ring buffer to be a `std::deque`.
Previously this was a `std::vector`. While we had reserved space for the
vector which would prevent growth and make `Enqueue` always cheap.
`Dequeue` ops would still require shifting the items in the vector.
`std::deque` gets rid of this limitation.

There is a tradeoff in memory usage here as `std::deque` will use more.
But current use cases only require small ring buffers. If we ever need
larger ring buffers, we should move to a more robust ring buffer
implementation that manages it's own contiguous memory.