Rework how valid rule IDs are tracked (#184)
The main purpose of this PR is to address a crash that could occur in
STANDALONE mode when attempting to approve ad hoc signed binaries.

This PR also changes behavior for how to identify which fields are valid
for rule lookup. Instead of clearing values form the SNTCachedDecision,
the signing status of the executing binary is tracked separately and
used to compute valid identifiers. This will allow reporting of all
fields in sync events instead of clearing this critical info (separate
PR).