Remove pending standalone notifications on UI connection invalidation (#213)
This change makes it so that not only are the pending blocks responded
to with a FALSE argument, but the notifications themselves are also
cleared from the queue.

Previously, the notifications were left in the queue to replicate
behavior similar to LOCKDOWN mode where the user would be notified of
blocks that occurred while not logged in and allow them to remedy the
situation.

However for standalone mode, the reply blocks were getting cleared, but
leaving the notification intact. This resulted in the user being
notified about the block event and given the option to approve the
binary. However since the reply block was cleared, santad would take no
action and no rule would be created. This would be confusing to users.

Some better potential long term fixes would be to add additional
AuthResultCache states to allow the replyBlock to be called after the
user eventually logged in. Another solution would be to present the
dialog to the user to inform them of the block but don't allow for
authorization. The user would still be able to perform manual
authorization (e.g. via santactl).