Idempotent config states (pt1) (#221)
This is part of a series of PRs that will move Santa to using
configuration "views" that provide a snapshot of configuration state
that can be used while processing some event or flow so that config
changes occur "atomically" and never partially apply during event
processing.

This first part sets up some initial infrastructure needed to support
this change. Namely it creates the new `SNTConfigState` object which
initially only captures `clientMode`. This also begins some of the work
of getting this object passed around to various components that will
need it in the future.

Finally, this PR addresses a UI-only issue that could result in the
wrong buttons being displayed if the client mode was changed to or from
Standalone mode via the sync server