Proc FAA - hookup probe and proc policy iteration interfaces (#244)
This PR sets the stage for the probe interface to do meaningful work by
hooking it up to the exec authorizer client as well as adding the
interfaces in WatchItems to support iterating policies. The "block
taking a block" design is to mitigate the need for the WatchItems class
to be aware of EndpointSecurity, and keeps policy matching in the FAA
clients.

Next PR will begin abstracting functions from the existing Data FAA
client for reuse in the Proc FAA client.