PathTargets cleanup, check policy block params, test migration (#272)
This change mostly moves tests from the Data FAA client to the new
FAAPolicyProcessor, along with some cleanup to work in the new class.
Other minor changes include changing the signature of
`CheckIfPolicyMatchesBlock` to prepare for allowing inverted Proc FAA
rules and removing and migrating away from the `PopulatePathTargets` to
the new, simpler, `PathTargets` interface.

Part of: #124