Add and adopt SantaSetCache (#291)
This PR removes the `ProcessSet` class from the Data FAA class and
replaces with a more abstract `SantaSetCache` built on top of the
existing `SantaCache`. The goal is to provide efficient, per-key set
operations where the operations on the underlying sets are performed
while under lock and are thread safe.

A new method, `contains` was added to `SantaCache` to allow for
thread-safe checking of values for caches with complex value types.

Additional abstractions were made to provide better encapsulation of
common code. Namely, a new `AuditUtilities` file was added to group
audittoken-specific functionality as well as add a new wrapper to
provide a pair of pid and pidversion.