Route Data FAA client through the FAAPolicyProcessor (#297)
This PR removes the now-duplicate code in the Data FAA client and routes
it through the same FAAPolicyProcessor as the Proc FAA client. While
this is mostly deleting code, some interface types had to be cleaned up
to conform to the new interfaces, requiring a large amount of touchups
in the tests, but the functionality didn't change.

This PR also results in the temporary loss of the following things that
will be added back in subsequent PRs:
* Rate limiting
* FAA event metrics

Additionally some outstanding work the is left for subsequent PRs to
keep this PR focused on the transition only:
* Renaming old FAA client to Data FAA
* Initializer updates to remove/replace items no longer needed or that
can now be simplified.