Handle FAA policy changes affecting watched processes (#311)
With this change, when the FAA config changes Santa will now re-evaluate
processes that are currently being watched because of the old policies
to ensure they should still be watched under the new policies.

This is done by simply clearing the rule cache and having policy lookup
be re-triggered the next time some event from a watched process comes
in. This also has the nice side effect that cache eviction is now
handled the same way.

This also fixes tracking/cleanup issues that existed on FORK/EXEC
events.