santad: use santa cache for event upload backoff (#341)
This CL replaces the NSCache used for event upload backoff with
SantaCache.

On full cache boundaries, NSCache, has exhibited some odd behavior.
Namely, it is possible to get into a situation where a hash is
successfully added to the cache, but is immediately cleared. Once in
this state, that hash is not able to be successfully retrieved. Even
after repeatedly being successfully added to the cache.

If an execution is being blocked in a tight loop like `while true; do
yes; done`, this could result in an event being uploaded every
[1500ms](https://github.com/northpolesec/santa/blob/82b09a99f3440a9eb6456a859a32905c6e2a5f90/Source/santad/EventProviders/AuthResultCache.h#L58).

Switching to SantaCache gives us better guarantees around full cache
boundaries. We know the cache will be cleared when overfilled, and we
know we can start filling it back up again. Once a usable LRU is
introduced to Santa's codebase, we could go that route, but it is not
pressing.



Also this CL adds some tests for SNTSyncdQueue.