Add config bundling (#346)
This PR introduces the "config bundles" which are a way to generically
refer to some config state that can be passed between components. This
is then used to pass settings between the sync service and main Santa
daemon.

Vacuuming the events database has been removed. The reason is that we
now vacuum on startup. Given it would be extremely rare for the database
to grow large, this only has the impact of adding database contention
unnecessarily. Additionally, constantly shrinking the database, even on
the "happy path", results in a performance penalty of having to keep
grow and shrink things.