santad: Fix 'UNIQUE constraint failed' and 'Unable to compute hash' errors (#368)
This fixes the following 2 errors, which are both appearing frequently
from santad:

```
default	11:38:26.799802-0400	com.northpolesec.santa.daemon	Unknown error calling sqlite3_step (19: UNIQUE constraint failed: events.filesha256) rs
```

```
default	11:38:27.456329-0400	com.northpolesec.santa.daemon	W com.northpolesec.santa.daemon: Unable to compute hash for critical system binary /System/Library/PrivateFrameworks/TCC.framework/Versions/A/Resources/tccd.
```

The first error is because of the unique constraint added in #337 - this
causes the transaction to fail if an event is added for a binary that
already has a pending event. The desired behavior is that the insert
failing is ignored, so adding `ON CONFLICT(filesha256) DO NOTHING` fixes
this. `INSERT OR IGNORE` would also work but that silently ignores *all*
constraints, including `NOT NULL`

The second error is because the location of `tccd` has moved from inside
the framework version to a support folder in some prior release of
macOS. Both paths are in the list of critical binaries but only 1 will
ever match. Logging these warnings could be confusing and the failure to
compute will never cause any problems.