docs: highlight santa profile content, update sizing (#381)
On the "Profile: Santa Configuration" page, highlight the part of the
profile that is _unique_ to Santa, to make it clearer where to put the
keys that we document.

While here, also update the container width so that wider windows don't
leave half the screen empty and update the "tab width" to 2 spaces so
that XML profiles can actually fit on a reasonable size screen.


![image](https://github.com/user-attachments/assets/671ac552-87d6-4faf-87cf-b036f702a472)