Collect, open, and send telemetry files to sync service for processing (#395)
The PR starts many of the foundational pieces for eventual support for
exporting telemetry.

The overall flow is:
1. Export is triggered via a timer in `santad`, or a new `santactl
telemetry` command
1.1. If triggered via `santactl`, an XPC message is sent to `santad` to
start exporting
2. `santad` will query the configured writer (based on the
`EventLogType`) for files to export
    1.1.  For each telemetry file, it is first opened
1.2. The file descriptor is sent to the sync service to perform the
actual export
* The file descriptor is passed because the sync service doesn't have
permissions to read the log files and also ensures logic for handling
log files is kept within the daemon
3. When `santad` receives a response from the sync service, it will
inform the writer whether or not the action was successful so it can do
appropriate cleanup.

This currently can only be enabled in debug builds, including the new
`santactl telemetry` command.