santactl: add eventupload command (#410)
This CL adds a `santactl eventupload` command. This allows a user to
upload metadata about an executable or bundle to a sync server, out of
band of Santa's typical operations.

There was a bit of refactoring work to make this possible, mainly around
making the bundle service multi-tenant. We don't want `santactl
eventupload` invocations interrupting each other or the GUI during
hashing events. In short, the bundle service is now safe to call from
multiple clients, allowing each request to be separately cancelable.


https://github.com/user-attachments/assets/2eb27e7d-56d1-4e30-bd2e-201d80ae9988