Keychain wrapper implementation (#425)
This adds objects for Santa to manage store/get/delete operations for
keychain items.

The `KeychainManager` is the main class. This will typically be used
like a singleton. For Santa's purposes, it will be injected when needed.
Consumers of the `KeychainManager` can create `KeychainItem`s for
themselves. The `KeychainItem` object then allows users to perform
keychain operations for that item.