common: Add CEL evaluator (#424)
This is the first part of implementing CEL rules (as a means to complete
#15, and google/santa#1200).

This PR adds the necessary dependencies and a new cel::Evaluator class
that holds a configured compiler. The evaluator can compile CEL
expressions into a re-usable expression plan and then evaluate it using
static per-binary data (e.g. signing timestamp) and blocks to retrieve
dynamic per-process data (e.g. command-line args).

This new library is as yet unused and won't be compiled in. The next PR
will add the ability to use CEL rules during execution.