gui: Fix minor UI issue when an app sets empty bundle name (#474)
This is a minor UI issue only that is rare and caused by an app
developer intentionally setting their bundle name to an empty string.
This issue only affects what data Santa chooses to show in the UI, and
could result in labels having missing values. There is no missing data
in sync events or telemetry.