Add Single Shot mode to Timer mixin class (#548)
Adds a new mode to the Timer mixin class called "SingleShot" that
derived classes can use. In this new mode, the timer is only fired once,
and then immediately cancelled while the OnTimer method is called. Upon
return, the timer is rescheduled.

This can be used by classes that might be performing long running
operations during the OnTimer callback and want to wait a full interval
before the operation is rescheduled.