Reenter temporary Monitor Mode on startup if time remaining (#639)
The PR adds Temp Monitor Mode state to the Santa state file. This allows
Santa to, on startup, check if there is any remaining time left in a
temporary Monitor Mode session and, if so, automatically reenter Monitor
Mode for the remainder of the previously authorized time.

The primary goal here is to allow authorized sessions to cross Santa
restarts (mainly due to background upgrades) so as not to confuse users.