Fix issue with bundle hash event URLs (#646)
This fixes an issue caused by #642 (this issue was never part of a
release)

The event detail URL template was expanded too early, prior to the
bundle hash being available. This meant that the link being opened to
the sync server had the main executable hash only and users would only
approve the main executable instead of the full app.