Refactor temporary monitor mode logic (#660)
This consolidates all the temporary monitor mode (TMM) logic into a
single class. This pulls code out of the daemon controller and
configurator. The configurator is still responsible for maintaining the
client mode state when in TMM and saving/maintaining the state file when
entering/leaving TMM - but it is not responsible for the
structure/contents of the TMM state. The daemon controller methods are
still needed as the entry point into TMM, but are essentially
pass-thrus. This consolidation will make the audit events much simpler.