Switch target field in FileAccess messages to be FileInfo type (#680)
The purpose of this change is to switch the `target` field in
`FileAccess` messages from `FileInfoLight` to `FileInfo`. This is an
on-wire compatible change that is also both forward and backward
compatible with the telemetry proto schemas since the message types
overlap and the additional fields in `FileInfo` are optional.

The result will be that for supported `FileAccess` events -
specifically, those events that have a real target file - stat
information will now be included. This will not always be the case for
all file access events. For example, FAA events related to attempts to
create a file will not have stat information because the resultant file
has not yet been created.