Remove message copies during Event Upload message creation (#692)
The previous code ended up performing an unnecessary copy due to
incorrect move semantics being used on the `std::optional` value. This
PR simplifies the code path to remove the optional altogether which
wasn't buying much. The copy is also removed by using
`UnsafeArenaAddAllocated`, which is safe in this instance since
everything is being created on the same arena.

Part of SNT-223